+++ src/esgd.rs
@@ -39,7 +39,7 @@ $$
 use candle_core::{Result, Var};
 use candle_nn::optim::Optimizer;
 
-use crate::{Decay, Momentum};
+use crate::{Decay, Momentum, OptimParams};
 
 /// Optimizer for Stochastic Gradient Descent with momentum.
 #[derive(Debug)]
@@ -55,7 +55,7 @@ struct VarSGD {
 }
 
 /// Parameters for SGD
-#[derive(Debug)]
+#[derive(Clone, Debug, PartialEq, PartialOrd)]
 pub struct ParamsSGD {
     /// Learning rate
     pub lr: f64,
@@ -344,6 +344,16 @@ impl Optimizer for SGD {
     }
 }
 
+impl OptimParams for SGD {
+    fn params(&self) -> &Self::Config {
+        &self.params
+    }
+
+    fn set_params(&mut self, config: Self::Config) {
+        self.params = config;
+    }
+}
+
 impl SGD {
     /// Return the vars being optimised
     #[must_use]
@@ -375,10 +385,10 @@ mod tests {
         // Now use backprop to run a linear regression between samples and get the coefficients back.
         let w = Var::new(&[[0f32, 0.]], &Device::Cpu)?;
         let b = Var::new(0f32, &Device::Cpu)?;
-        let mut n_sgd = SGD::new(vec![w.clone(), b.clone()], params)?;
-        assert_approx_eq!(0.004, n_sgd.learning_rate());
-        n_sgd.set_learning_rate(0.002);
-        assert_approx_eq!(0.002, n_sgd.learning_rate());
+        let mut optim = SGD::new(vec![w.clone(), b.clone()], params)?;
+        assert_approx_eq!(0.004, optim.learning_rate());
+        optim.set_learning_rate(0.002);
+        assert_approx_eq!(0.002, optim.learning_rate());
         Ok(())
     }
 
@@ -387,10 +397,30 @@ mod tests {
         let params = ParamsSGD::default();
         let w = Var::new(&[[3f32, 1.]], &Device::Cpu)?;
         let b = Var::new(-2f32, &Device::Cpu)?;
-        let n_sgd = SGD::new(vec![w.clone(), b.clone()], params)?;
-        let inner = n_sgd.into_inner();
+        let optim = SGD::new(vec![w.clone(), b.clone()], params)?;
+        let inner = optim.into_inner();
         assert_eq!(inner[0].as_tensor().to_vec2::<f32>()?, &[[3f32, 1.]]);
         assert_approx_eq!(inner[1].as_tensor().to_vec0::<f32>()?, -2_f32);
+        Ok(())
+    }
+
+    #[test]
+    fn params_test() -> Result<()> {
+        let params = ParamsSGD {
+            lr: 0.004,
+            ..Default::default()
+        };
+        // Now use backprop to run a linear regression between samples and get the coefficients back.
+        let w = Var::new(&[[0f32, 0.]], &Device::Cpu)?;
+        let b = Var::new(0f32, &Device::Cpu)?;
+        let mut optim = SGD::new(vec![w.clone(), b.clone()], params.clone())?;
+        assert_eq!(params, optim.params().clone());
+        let new_params = ParamsSGD {
+            lr: 0.002,
+            ..Default::default()
+        };
+        optim.set_params(new_params.clone());
+        assert_eq!(new_params, optim.params().clone());
         Ok(())
     }
 }