+++ src/lib.rs
@@ -18,6 +18,14 @@ pub mod nadam;
 pub mod radam;
 pub mod rmsprop;
 
+/// Trait for optimisers to expose their parameters
+pub trait OptimParams: candle_nn::optim::Optimizer {
+    /// get the current parameters of the Optimiser
+    fn params(&self) -> &Self::Config;
+    /// set the current parameters of the Optimiser
+    fn set_params(&mut self, config: Self::Config);
+}
+
 /// Trait for Models: this is needed for optimisers that require the ability to calculate the loss
 /// such as LBFGS
 pub trait Model: Sized {
@@ -59,7 +67,7 @@ pub enum ModelOutcome {
 }
 
 /// Method of weight decay to use
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, PartialEq, PartialOrd)]
 pub enum Decay {
     /// Weight decay regularisation to penalise large weights
     ///
@@ -80,7 +88,7 @@ pub enum Decay {
 }
 
 /// Type of momentum to use
-#[derive(Copy, Clone, Debug)]
+#[derive(Copy, Clone, Debug, PartialEq, PartialOrd)]
 pub enum Momentum {
     /// classical momentum
     Classical(f64),